requirements.txt
Django==5.2.4
djangorestframework==3.16.0
Pillow==11.3.0
python-decouple==3.8
django-cors-headers==4.7.0
gunicorn==20.1.0
psycopg2-binary==2.9.7